# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScopeGuard API specification - change order approval workflow for agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/me:
    get:
      operationId: getMyOrganization
      tags: [organizations]
      summary: Get current user's organization
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateMyOrganization
      tags: [organizations]
      summary: Update current user's organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  # CLIENT CONTACTS
  /client-contacts:
    get:
      operationId: listClientContacts
      tags: [client-contacts]
      summary: List all client contacts for current org
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of client contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientContact"
    post:
      operationId: createClientContact
      tags: [client-contacts]
      summary: Create a new client contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientContactBody"
      responses:
        "201":
          description: Created client contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientContact"
        "409":
          description: Contact with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /client-contacts/{id}:
    get:
      operationId: getClientContact
      tags: [client-contacts]
      summary: Get a single client contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientContact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClientContact
      tags: [client-contacts]
      summary: Update a client contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientContactBody"
      responses:
        "200":
          description: Updated client contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientContact"
    delete:
      operationId: deleteClientContact
      tags: [client-contacts]
      summary: Delete a client contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # PROJECTS
...
  /change-requests:
    get:
      operationId: listChangeRequests
      tags: [change-requests]
      summary: List all change requests for current org
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SENT, APPROVED, REJECTED, REVISED]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of change requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeRequest"
    post:
      operationId: createChangeRequest
      tags: [change-requests]
      summary: Create a new change request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChangeRequestBody"
      responses:
        "201":
          description: Created change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeRequest"
  /change-requests/{id}:
    get:
      operationId: getChangeRequest
      tags: [change-requests]
      summary: Get a single change request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateChangeRequest
      tags: [change-requests]
      summary: Update a change request (only DRAFT/REVISED)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChangeRequestBody"
      responses:
        "200":
          description: Updated change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeRequest"
    delete:
      operationId: deleteChangeRequest
      tags: [change-requests]
      summary: Delete a change request (only DRAFT)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /change-requests/{id}/send:
    post:
      operationId: sendChangeRequest
      tags: [change-requests]
      summary: Send a change request to the client (transitions to SENT)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Change request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeRequest"
        "400":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PUBLIC APPROVAL (no auth required)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]

    # AUTH SCHEMAS
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt]
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true

    # CLIENT CONTACT SCHEMAS
    ClientContact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        company:
          type: string
        organizationId:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, company, organizationId, createdAt]
    CreateClientContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
      required: [name, email]
    UpdateClientContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string

    # PROJECT SCHEMAS
...
    ChangeRequest:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
        hourlyRateCents:
          type: integer
        totalCents:
          type: integer
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, REVISED]
        deadline:
          type: string
          format: date
          nullable: true
        approvalToken:
          type: string
        approvedAt:
          type: string
          format: date-time
          nullable: true
        clientComments:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, projectName, clientName, clientEmail, title, description, estimatedHours, hourlyRateCents, totalCents, status, approvalToken, clientComments, createdAt, updatedAt]
    CreateChangeRequestBody:
      type: object
      properties:
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
          minimum: 0
        deadline:
          type: string
          format: date
          nullable: true
      required: [projectId, title, description, estimatedHours]
    UpdateChangeRequestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
          minimum: 0
        deadline:
          type: string
          format: date
          nullable: true

    # PUBLIC APPROVAL SCHEMAS
...